requirements.txt
aniso8601==7.0.0
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==2.1.5
django-cors-headers==3.2.1
django-cursor-pagination==0.1.4
django-elasticsearch-dsl==6.4.1
django-filter==2.2.0
django-graphql-jwt==0.3.1
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
elasticsearch==6.8.1
elasticsearch-dsl==6.4.0
google==2.0.3
google-api-core==1.16.0
google-auth==1.12.0
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.9.1
graphql-core==2.3.1
graphql-relay==2.0.1
grpcio==1.27.2
gunicorn==20.0.4
idna==2.9
Pillow==7.1.1
promise==2.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
